+++ Sources/KeyboardKit/Actions/StandardKeyboardActionHandler.swift
@@ -145,7 +145,7 @@ open class StandardKeyboardActionHandler: NSObject, KeyboardActionHandler {
         guard let gestureAction = self.action(for: gesture, on: action) else { return }
         tryRemoveAutocompleteInsertedSpace(before: gesture, on: action)
         tryApplyAutocompleteSuggestion(before: gesture, on: action)
-        gestureAction(.shared)
+        gestureAction(KeyboardInputViewController.shared)
         tryReinsertAutocompleteRemovedSpace(after: gesture, on: action)
         tryEndSentence(after: gesture, on: action)
         tryChangeKeyboardType(after: gesture, on: action)